Generalize interface, simplify the extension to more methods
